refactor: keep plugin command dispatch in CLI mixin
