+++ cli.py
@@ -3296,27 +3296,6 @@ class HermesCLI(CLIProcessNotificationsMixin, CLIAgentSetupMixin, CLICommandsMix
             self._console_print(f"[bold red]Quick command error: {e}[/]")
         return True
 
-    def _run_plugin_slash_command(self, base_cmd: str, user_args: str) -> None:
-        from hermes_cli.plugins import get_plugin_command_handler, invoke_plugin_command, resolve_plugin_command_result
-
-        plugin_handler = get_plugin_command_handler(base_cmd.lstrip("/"))
-        if not plugin_handler:
-            return
-        try:
-            result = resolve_plugin_command_result(
-                invoke_plugin_command(
-                    plugin_handler,
-                    user_args,
-                    session_id=getattr(self, "session_id", None),
-                    session_key=getattr(self, "session_id", None),
-                    platform="cli",
-                )
-            )
-            if result:
-                _cprint(str(result))
-        except Exception as e:
-            _cprint(f"\033[1;31mPlugin command error: {e}{_RST}")
-
     def _queue_skill_message(self, msg) -> None:
         if hasattr(self, '_pending_input'):
             self._pending_input.put(msg)
