+++ hermes_cli/cli_commands_mixin.py
@@ -602,6 +602,29 @@ def _browser_status() -> None:
 class CLICommandsMixin:
     """Mixin holding the interactive-CLI slash-command handlers."""
 
+    def _run_plugin_slash_command(self, base_cmd: str, user_args: str) -> None:
+        """Run a discovered plugin slash command with the current CLI session context."""
+        from cli import _RST, _cprint
+        from hermes_cli.plugins import get_plugin_command_handler, invoke_plugin_command, resolve_plugin_command_result
+
+        plugin_handler = get_plugin_command_handler(base_cmd.lstrip("/"))
+        if not plugin_handler:
+            return
+        try:
+            result = resolve_plugin_command_result(
+                invoke_plugin_command(
+                    plugin_handler,
+                    user_args,
+                    session_id=getattr(self, "session_id", None),
+                    session_key=getattr(self, "session_id", None),
+                    platform="cli",
+                )
+            )
+            if result:
+                _cprint(str(result))
+        except Exception as exc:
+            _cprint(f"\033[1;31mPlugin command error: {exc}{_RST}")
+
     # ---- /rollback ------------------------------------------------------------------------
     def _checkpoint_manager(self, disabled_lines):
         """The agent's checkpoint manager, or None after printing why it is unavailable."""
